requirements.txt
python-telegram-bot==21.5
google-generativeai>=0.8.0